Cast everything to a float when writing sqlite db
It turns out that if you load an sqlite db with sqldf in R, and there
are no column types, it uses the first value to determine the type in
the data frame.

If that value is an integer, and the rest of the values are floats…
Then they all get turned into ints and things go horribly and opaquely
wrong.

Easiest solution: cast everything to a float unless it is a string, or
the cast is for some reason impossible.